app/consumer/requirements.txt
kafka-python==2.1.5
psycopg2-binary==2.9.10
